refactor: localStorage 에서 값을 가져와 초기화하고, 처음 화면 보여주는 기능 외부 모듈로 분리
